refactor(FormNavigationButtons): Convert to TypeScript
